feat(layout): improve tabs (#36294)
