+++ frontend/src/lib/components/CommandBar/CommandBar.stories.tsx
@@ -6,7 +6,7 @@ import { BarStatus } from 'lib/components/CommandBar/types'
 import { mswDecorator } from '~/mocks/browser'
 
 import { CommandBar } from './CommandBar'
-import { useOnMountEffect } from 'lib/hooks/useOnMountEffect'
+import { useDelayedOnMountEffect } from 'lib/hooks/useOnMountEffect'
 
 const SEARCH_RESULT = {
     results: [
@@ -536,21 +536,21 @@ export default meta
 
 export function Search(): JSX.Element {
     const { setCommandBar } = useActions(commandBarLogic)
-    useOnMountEffect(() => setCommandBar(BarStatus.SHOW_SEARCH))
+    useDelayedOnMountEffect(() => setCommandBar(BarStatus.SHOW_SEARCH))
 
     return <CommandBar />
 }
 
 export function Actions(): JSX.Element {
     const { setCommandBar } = useActions(commandBarLogic)
-    useOnMountEffect(() => setCommandBar(BarStatus.SHOW_ACTIONS))
+    useDelayedOnMountEffect(() => setCommandBar(BarStatus.SHOW_ACTIONS))
 
     return <CommandBar />
 }
 
 export function Shortcuts(): JSX.Element {
     const { setCommandBar } = useActions(commandBarLogic)
-    useOnMountEffect(() => setCommandBar(BarStatus.SHOW_SHORTCUTS))
+    useDelayedOnMountEffect(() => setCommandBar(BarStatus.SHOW_SHORTCUTS))
 
     return <CommandBar />
 }
