+++ frontend/src/lib/components/TaxonomicFilter/TaxonomicFilter.stories.tsx
@@ -9,7 +9,7 @@ import { AvailableFeature } from '~/types'
 
 import { infiniteListLogic } from './infiniteListLogic'
 import { TaxonomicFilter } from './TaxonomicFilter'
-import { useOnMountEffect } from 'lib/hooks/useOnMountEffect'
+import { useDelayedOnMountEffect } from 'lib/hooks/useOnMountEffect'
 
 const meta: Meta<typeof TaxonomicFilter> = {
     title: 'Filters/Taxonomic Filter',
@@ -41,7 +41,7 @@ export const EventsFree: StoryFn<typeof TaxonomicFilter> = (args) => {
 
     // Highlight the second item, as the first one is "All events", which doesn't have a definition to show
     // - we do want to show the definition popover here too
-    useOnMountEffect(() => setIndex(1))
+    useDelayedOnMountEffect(() => setIndex(1))
 
     return (
         <div className="w-fit border rounded p-2 bg-surface-primary">
@@ -90,7 +90,7 @@ export const Actions: StoryFn<typeof TaxonomicFilter> = (args) => {
 
     // Highlight the second item, as the first one is "All events", which doesn't have a definition to show
     // - we do want to show the definition popover here too
-    useOnMountEffect(() => setIndex(0))
+    useDelayedOnMountEffect(() => setIndex(0))
 
     return (
         <div className="w-fit border rounded p-2">
@@ -144,7 +144,7 @@ export const Columnar: StoryFn<typeof TaxonomicFilter> = (args) => {
         })
     )
 
-    useOnMountEffect(() => setIndex(1))
+    useDelayedOnMountEffect(() => setIndex(1))
 
     return (
         <div className="w-fit border rounded p-2 bg-surface-primary">
@@ -185,7 +185,7 @@ export const ForceColumnar: StoryFn<typeof TaxonomicFilter> = (args) => {
         })
     )
 
-    useOnMountEffect(() => setIndex(1))
+    useDelayedOnMountEffect(() => setIndex(1))
 
     return (
         <div className="w-fit border rounded p-2 bg-surface-primary">
@@ -225,7 +225,7 @@ export const ForceNonColumnar: StoryFn<typeof TaxonomicFilter> = (args) => {
         })
     )
 
-    useOnMountEffect(() => setIndex(1))
+    useDelayedOnMountEffect(() => setIndex(1))
 
     return (
         <div className="w-fit border rounded p-2 bg-surface-primary">
