+++ frontend/src/lib/hooks/useOnMountEffect.ts
@@ -33,3 +33,9 @@ import React, { useEffect } from 'react'
 export function useOnMountEffect(effect: React.EffectCallback): void {
     useEffect(effect, []) // oxlint-disable-line react-hooks/exhaustive-deps
 }
+
+export function useDelayedOnMountEffect(effect: React.EffectCallback, timeout = 500): void {
+    useEffect(() => {
+        window.setTimeout(effect, timeout)
+    }, []) // oxlint-disable-line react-hooks/exhaustive-deps
+}
