+++ frontend/src/scenes/authentication/signup/verify-email/VerifyEmail.stories.tsx
@@ -2,7 +2,7 @@ import { Meta, Story } from '@storybook/react'
 
 import { VerifyEmail } from './VerifyEmail'
 import { verifyEmailLogic } from './verifyEmailLogic'
-import { useOnMountEffect } from 'lib/hooks/useOnMountEffect'
+import { useDelayedOnMountEffect } from 'lib/hooks/useOnMountEffect'
 
 const meta: Meta = {
     title: 'Scenes-Other/Verify Email',
@@ -13,7 +13,7 @@ const meta: Meta = {
 }
 export default meta
 export const VerifyEmailPending: Story = () => {
-    useOnMountEffect(() => {
+    useDelayedOnMountEffect(() => {
         verifyEmailLogic.actions.setView('pending')
         verifyEmailLogic.actions.setUuid('12345678')
     })
@@ -22,7 +22,7 @@ export const VerifyEmailPending: Story = () => {
 }
 
 export const VerifyingEmail: Story = () => {
-    useOnMountEffect(() => {
+    useDelayedOnMountEffect(() => {
         verifyEmailLogic.actions.setView('verify')
     })
 
@@ -35,7 +35,7 @@ VerifyingEmail.parameters = {
 }
 
 export const VerifyEmailSuccess: Story = () => {
-    useOnMountEffect(() => {
+    useDelayedOnMountEffect(() => {
         verifyEmailLogic.actions.setView('success')
     })
 
@@ -43,7 +43,7 @@ export const VerifyEmailSuccess: Story = () => {
 }
 
 export const VerifyEmailInvalid: Story = () => {
-    useOnMountEffect(() => {
+    useDelayedOnMountEffect(() => {
         verifyEmailLogic.actions.setView('invalid')
         verifyEmailLogic.actions.setUuid('12345678')
     })
