+++ frontend/src/scenes/insights/EmptyStates/EmptyStates.stories.tsx
@@ -8,7 +8,7 @@ import { InsightShortId } from '~/types'
 import insight from '../../../mocks/fixtures/api/projects/team_id/insights/trendsLine.json'
 import { insightVizDataLogic } from '../insightVizDataLogic'
 import funnelOneStep from './funnelOneStep.json'
-import { useOnMountEffect } from 'lib/hooks/useOnMountEffect'
+import { useDelayedOnMountEffect } from 'lib/hooks/useOnMountEffect'
 
 type Story = StoryObj<typeof App>
 const meta: Meta = {
@@ -159,7 +159,7 @@ export const LongLoading: StoryFn = () => {
         },
     })
 
-    useOnMountEffect(() => {
+    useDelayedOnMountEffect(() => {
         const logic = insightVizDataLogic.findMounted({ dashboardItemId: insight.short_id as InsightShortId })
         logic?.actions.setTimedOutQueryId('a-uuid-query-id') // Show the suggestions immediately
     })
