+++ frontend/src/scenes/pipeline/Pipeline.stories.tsx
@@ -14,7 +14,7 @@ import pluginConfigs from './__mocks__/pluginConfigs.json'
 import plugins from './__mocks__/plugins.json'
 import { appsManagementLogic } from './appsManagementLogic'
 import { pipelineNodeMetricsLogic } from './pipelineNodeMetricsLogic'
-import { useOnMountEffect } from 'lib/hooks/useOnMountEffect'
+import { useDelayedOnMountEffect } from 'lib/hooks/useOnMountEffect'
 
 const pluginRetrieveMock: MockSignature = (req, res, ctx) => {
     const plugin = plugins.results.find((conf) => conf.id === Number(req.params.id))
@@ -207,7 +207,7 @@ export const PipelineNodeConfiguration404: Story = {
 }
 
 export function PipelineNodeMetrics(): JSX.Element {
-    useOnMountEffect(() => pipelineNodeMetricsLogic({ id: geoIpConfigId }).mount())
+    useDelayedOnMountEffect(() => pipelineNodeMetricsLogic({ id: geoIpConfigId }).mount())
 
     return <App />
 }
@@ -216,7 +216,7 @@ PipelineNodeMetrics.parameters = {
 }
 
 export function PipelineNodeMetricsErrorModal(): JSX.Element {
-    useOnMountEffect(() => {
+    useDelayedOnMountEffect(() => {
         const logic = pipelineNodeMetricsLogic({ id: geoIpConfigId })
         logic.mount()
         logic.actions.openErrorDetailsModal('Error')
@@ -253,7 +253,7 @@ export const PipelineNodeLogsBatchExport: Story = {
 }
 
 export function PipelineNodesManagementPage(): JSX.Element {
-    useOnMountEffect(() => appsManagementLogic.mount())
+    useDelayedOnMountEffect(() => appsManagementLogic.mount())
 
     return <App />
 }
