+++ frontend/src/scenes/sceneLogic.ts
@@ -1,10 +1,10 @@
-import { actions, BuiltLogic, connect, kea, listeners, path, props, reducers, selectors } from 'kea'
-import { router, urlToAction } from 'kea-router'
+import { actions, afterMount, BuiltLogic, connect, kea, listeners, path, props, reducers, selectors } from 'kea'
+import { combineUrl, router, urlToAction } from 'kea-router'
 import { commandBarLogic } from 'lib/components/CommandBar/commandBarLogic'
 import { BarStatus } from 'lib/components/CommandBar/types'
 import { TeamMembershipLevel } from 'lib/constants'
-import { getRelativeNextPath } from 'lib/utils'
-import { removeProjectIdIfPresent } from 'lib/utils/router-utils'
+import { identifierToHuman, getRelativeNextPath } from 'lib/utils'
+import { addProjectIdIfMissing, removeProjectIdIfPresent } from 'lib/utils/router-utils'
 import { withForwardedSearchParams } from 'lib/utils/sceneLogicUtils'
 import posthog from 'posthog-js'
 import {
@@ -36,6 +36,34 @@ import type { sceneLogicType } from './sceneLogicType'
 import { inviteLogic } from './settings/organization/inviteLogic'
 import { teamLogic } from './teamLogic'
 import { userLogic } from './userLogic'
+import { arrayMove } from '@dnd-kit/sortable'
+import { subscriptions } from 'kea-subscriptions'
+
+export interface SceneTab {
+    id: string
+    pathname: string
+    search: string
+    hash: string
+    title: string
+    active: boolean
+}
+
+const TAB_STATE_KEY = 'scene-tabs-state'
+const persistTabs = (tabs: SceneTab[]): void => {
+    sessionStorage.setItem(TAB_STATE_KEY, JSON.stringify(tabs))
+}
+const getPersistedTabs: () => SceneTab[] | null = () => {
+    const savedTabs = sessionStorage.getItem(TAB_STATE_KEY)
+    if (savedTabs) {
+        try {
+            return JSON.parse(savedTabs)
+        } catch (e) {
+            console.error('Failed to parse saved tabs from sessionStorage:', e)
+        }
+    }
+    return null
+}
+const generateTabId = (): string => crypto?.randomUUID?.() || `${Date.now()}-${Math.random()}`
 
 export const productUrlMapping: Partial<Record<ProductKey, string[]>> = {
     [ProductKey.SESSION_REPLAY]: [urls.replay()],
@@ -72,9 +100,19 @@ export const sceneLogic = kea<sceneLogicType>([
     path(['scenes', 'sceneLogic']),
     connect(() => ({
         logic: [router, userLogic, preflightLogic],
-        actions: [router, ['locationChanged'], commandBarLogic, ['setCommandBar'], inviteLogic, ['hideInviteModal']],
+        actions: [
+            router,
+            ['locationChanged', 'push'],
+            commandBarLogic,
+            ['setCommandBar'],
+            inviteLogic,
+            ['hideInviteModal'],
+        ],
         values: [billingLogic, ['billing'], organizationLogic, ['organizationBeingDeleted']],
     })),
+    afterMount(({ cache }) => {
+        cache.mountedTabLogic = {} as Record<string, () => void>
+    }),
     actions({
         /* 1. Prepares to open the scene, as the listener may override and do something
         else (e.g. redirecting if unauthenticated), then calls (2) `loadScene`*/
@@ -109,6 +147,13 @@ export const sceneLogic = kea<sceneLogicType>([
             loadedScene,
         }),
         reloadBrowserDueToImportError: true,
+
+        newTab: true,
+        setTabs: (tabs: SceneTab[]) => ({ tabs }),
+        removeTab: (tab: SceneTab) => ({ tab }),
+        activateTab: (tab: SceneTab) => ({ tab }),
+        clickOnTab: (tab: SceneTab) => ({ tab }),
+        reorderTabs: (activeId: string, overId: string) => ({ activeId, overId }),
     }),
     reducers({
         scene: [
@@ -151,6 +196,61 @@ export const sceneLogic = kea<sceneLogicType>([
             { persist: true },
             {
                 reloadBrowserDueToImportError: () => new Date().valueOf(),
+            },
+        ],
+        tabs: [
+            [] as SceneTab[],
+            {
+                setTabs: (_, { tabs }) => tabs,
+                newTab: (state) => {
+                    return [
+                        ...state.map((tab) => (tab.active ? { ...tab, active: false } : tab)),
+                        {
+                            id: generateTabId(),
+                            active: true,
+                            pathname: addProjectIdIfMissing('/new'),
+                            search: '',
+                            hash: '',
+                            title: 'New tab',
+                        },
+                    ]
+                },
+                removeTab: (state, { tab }) => {
+                    let index = state.findIndex((t) => t === tab)
+                    if (index === -1) {
+                        console.error('Tab to remove not found', tab)
+                        return state
+                    }
+                    let newState = state.filter((_, i) => i !== index)
+                    if (!newState.find((t) => t.active)) {
+                        const newActiveIndex = Math.max(index - 1, 0)
+                        newState = newState.map((tab, i) => (i === newActiveIndex ? { ...tab, active: true } : tab))
+                    }
+                    return newState
+                },
+                activateTab: (state, { tab }) => {
+                    const newState = state.map((t) =>
+                        t === tab
+                            ? !t.active
+                                ? { ...t, active: true }
+                                : t
+                            : t.active
+                              ? {
+                                    ...t,
+                                    active: false,
+                                }
+                              : t
+                    )
+                    return newState
+                },
+                reorderTabs: (state, { activeId, overId }) => {
+                    const oldIndex = state.findIndex((t) => t.id === activeId)
+                    const newIndex = state.findIndex((t) => t.id === overId)
+                    if (oldIndex === -1 || newIndex === -1 || oldIndex === newIndex) {
+                        return state
+                    }
+                    return arrayMove(state, oldIndex, newIndex)
+                },
             },
         ],
     }),
@@ -196,11 +296,15 @@ export const sceneLogic = kea<sceneLogicType>([
                 activeLoadedScene?.sceneParams || { params: {}, searchParams: {}, hashParams: {} },
         ],
         activeSceneLogic: [
-            (s) => [s.activeLoadedScene, s.sceneParams],
-            (activeLoadedScene, sceneParams): BuiltLogic | null =>
-                activeLoadedScene?.logic
-                    ? activeLoadedScene.logic.build(activeLoadedScene.paramsToProps?.(sceneParams) || {})
-                    : null,
+            (s) => [s.activeLoadedScene, s.sceneParams, s.activeTabId],
+            (activeLoadedScene, sceneParams, activeTabId): BuiltLogic | null => {
+                return activeLoadedScene?.logic
+                    ? activeLoadedScene.logic.build({
+                          ...activeLoadedScene.paramsToProps?.(sceneParams),
+                          tabId: activeTabId,
+                      })
+                    : null
+            },
         ],
         params: [(s) => [s.sceneParams], (sceneParams): Record<string, string> => sceneParams.params || {}],
         searchParams: [(s) => [s.sceneParams], (sceneParams): Record<string, any> => sceneParams.searchParams || {}],
@@ -217,15 +321,184 @@ export const sceneLogic = kea<sceneLogicType>([
                 return null
             },
         ],
+        activeTab: [
+            (s) => [s.tabs],
+            (tabs: SceneTab[]): SceneTab | null => {
+                return tabs.find((tab) => tab.active) || null
+            },
+        ],
+        activeTabId: [
+            (s) => [s.activeTab],
+            (activeTab: SceneTab | null): string | null => {
+                return activeTab ? activeTab.id : null
+            },
+        ],
+        tabIds: [
+            (s) => [s.tabs],
+            (tabs: SceneTab[]): Record<string, boolean> => {
+                return tabs.reduce(
+                    (acc, tab) => {
+                        acc[tab.id] = true
+                        return acc
+                    },
+                    {} as Record<string, boolean>
+                )
+            },
+        ],
+        title: [
+            (s) => [
+                // We're effectively passing the selector through to the scene logic, and "recalculating"
+                // this every time it's rendered. Caching will happen within the scene's breadcrumb selector.
+                (state, props): string => {
+                    const activeSceneLogic = sceneLogic.selectors.activeSceneLogic(state, props)
+                    if (activeSceneLogic && 'breadcrumbs' in activeSceneLogic.selectors) {
+                        try {
+                            const activeLoadedScene = sceneLogic.selectors.activeLoadedScene(state, props)
+                            const bc = activeSceneLogic.selectors.breadcrumbs(
+                                state,
+                                activeLoadedScene?.paramsToProps?.(activeLoadedScene?.sceneParams) || props
+                            )
+                            return bc.length > 0 ? bc[bc.length - 1].name : '...'
+                        } catch {
+                            // If the breadcrumb selector fails, we'll just ignore it and return a placeholder value below
+                        }
+                    }
+
+                    const activeScene = s.activeScene(state, props)
+                    if (activeScene) {
+                        const sceneConfig = s.sceneConfig(state, props)
+                        return sceneConfig?.name ?? identifierToHuman(activeScene)
+                    }
+                    return '...'
+                },
+            ],
+            (title): string => title,
+        ],
     }),
-    listeners(({ values, actions, props, selectors }) => ({
-        setScene: ({ scene, scrollToTop }, _, __, previousState) => {
+    listeners(({ values, actions, cache, props, selectors }) => ({
+        newTab: () => {
+            persistTabs(values.tabs)
+            router.actions.push(urls.newTab())
+        },
+        setTabs: () => persistTabs(values.tabs),
+        activateTab: () => persistTabs(values.tabs),
+        removeTab: ({ tab }) => {
+            if (tab.active) {
+                // values.activeTab will already be the new active tab from the reducer
+                const { activeTab } = values
+                if (activeTab) {
+                    router.actions.push(activeTab.pathname, activeTab.search, activeTab.hash)
+                } else {
+                    persistTabs(values.tabs)
+                }
+            } else {
+                persistTabs(values.tabs)
+            }
+        },
+        clickOnTab: ({ tab }) => {
+            if (!tab.active) {
+                actions.activateTab(tab)
+            }
+            router.actions.push(tab.pathname, tab.search, tab.hash)
+            persistTabs(values.tabs)
+        },
+        reorderTabs: () => {
+            persistTabs(values.tabs)
+        },
+        push: ({ url, hashInput, searchInput }) => {
+            let { pathname, search, hash } = combineUrl(url, searchInput, hashInput)
+            pathname = addProjectIdIfMissing(pathname)
+
+            const activeTabIndex = values.tabs.findIndex((tab) => tab.active)
+            if (activeTabIndex !== -1) {
+                const newTabs = values.tabs.map((tab, i) =>
+                    i === activeTabIndex
+                        ? { ...tab, active: true, pathname, search, hash }
+                        : tab.active
+                          ? {
+                                ...tab,
+                                active: false,
+                            }
+                          : tab
+                )
+                actions.setTabs(newTabs)
+            } else {
+                actions.setTabs([
+                    ...values.tabs,
+                    { id: generateTabId(), active: true, pathname, search, hash, title: 'Loading...' },
+                ])
+            }
+            persistTabs(values.tabs)
+        },
+        locationChanged: ({ pathname, search, hash, routerState, method }) => {
+            pathname = addProjectIdIfMissing(pathname)
+            if (routerState?.tabs && method !== 'PUSH') {
+                actions.setTabs(routerState.tabs)
+                return
+            }
+            const activeTabIndex = values.tabs.findIndex((tab) => tab.active)
+            if (activeTabIndex !== -1) {
+                actions.setTabs(
+                    values.tabs.map((tab, i) =>
+                        i === activeTabIndex
+                            ? { ...tab, active: true, pathname, search, hash }
+                            : tab.active
+                              ? {
+                                    ...tab,
+                                    active: false,
+                                }
+                              : tab
+                    )
+                )
+            } else {
+                actions.setTabs([
+                    ...values.tabs,
+                    { id: generateTabId(), active: true, pathname, search, hash, title: 'Loading...' },
+                ])
+            }
+            persistTabs(values.tabs)
+
+            // Open search or command bar
+            const params = new URLSearchParams(search)
+            const searchBar = params.get('searchBar')
+            const commandBar = params.get('commandBar')
+
+            if (searchBar !== null) {
+                actions.setCommandBar(BarStatus.SHOW_SEARCH, searchBar)
+                params.delete('searchBar')
+                router.actions.replace(pathname, params, hash)
+            } else if (commandBar !== null) {
+                actions.setCommandBar(BarStatus.SHOW_ACTIONS, commandBar)
+                params.delete('commandBar')
+                router.actions.replace(pathname, params, hash)
+            }
+
+            // Remove trailing slash
+            if (pathname !== '/' && pathname.endsWith('/')) {
+                router.actions.replace(pathname.replace(/(\/+)$/, ''), search, hash)
+            }
+        },
+        setScene: ({ scene, loadedScene, params, scrollToTop }, _, __, previousState) => {
             posthog.capture('$pageview')
 
             // if we clicked on a link, scroll to top
             const previousScene = selectors.scene(previousState)
             if (scrollToTop && scene !== previousScene) {
                 window.scrollTo(0, 0)
+            }
+
+            const activeTabId = values.activeTabId
+            if (!activeTabId) {
+                console.warn('No active tab found when setting scene logic')
+                return
+            }
+            const unmount = cache.mountedTabLogic[activeTabId]
+            if (unmount) {
+                window.setTimeout(unmount, 50)
+                delete cache.mountedTabLogic[activeTabId]
+            }
+            if (loadedScene?.logic) {
+                cache.mountedTabLogic[activeTabId] = loadedScene?.logic(loadedScene?.paramsToProps?.(params)).mount()
             }
         },
         openScene: ({ scene, sceneKey, params, method }) => {
@@ -385,7 +658,9 @@ export const sceneLogic = kea<sceneLogicType>([
                 } finally {
                     window.clearTimeout(timeout)
                 }
-                breakpoint()
+                if (values.scene !== scene) {
+                    breakpoint()
+                }
                 const { default: defaultExport, logic, scene: _scene, ...others } = importedScene
 
                 if (_scene) {
@@ -432,32 +707,33 @@ export const sceneLogic = kea<sceneLogicType>([
         reloadBrowserDueToImportError: () => {
             window.location.reload()
         },
-        locationChanged: () => {
-            const {
-                location: { pathname, search, hash },
-            } = router.values
-
-            // Open search or command bar
-            const params = new URLSearchParams(search)
-            const searchBar = params.get('searchBar')
-            const commandBar = params.get('commandBar')
-
-            if (searchBar !== null) {
-                actions.setCommandBar(BarStatus.SHOW_SEARCH, searchBar)
-                params.delete('searchBar')
-                router.actions.replace(pathname, params, hash)
-            } else if (commandBar !== null) {
-                actions.setCommandBar(BarStatus.SHOW_ACTIONS, commandBar)
-                params.delete('commandBar')
-                router.actions.replace(pathname, params, hash)
-            }
-
-            // Remove trailing slash
-            if (pathname !== '/' && pathname.endsWith('/')) {
-                router.actions.replace(pathname.replace(/(\/+)$/, ''), search, hash)
-            }
-        },
     })),
+
+    // keep this above subscriptions
+    afterMount(({ actions, cache, values }) => {
+        if (!cache.tabsLoaded) {
+            const savedTabs = getPersistedTabs()
+            const withIds = savedTabs?.map((t) => (t.id ? t : { ...t, id: generateTabId() }))
+            if (withIds) {
+                actions.setTabs(withIds)
+            }
+            cache.tabsLoaded = true
+        }
+        if (values.tabs.length === 0) {
+            const { currentLocation } = router.values
+            actions.setTabs([
+                {
+                    id: generateTabId(),
+                    active: true,
+                    pathname: currentLocation.pathname,
+                    search: currentLocation.search,
+                    hash: currentLocation.hash,
+                    title: 'Loading...',
+                },
+            ])
+        }
+    }),
+
     urlToAction(({ actions }) => {
         const mapping: Record<
             string,
@@ -493,4 +769,49 @@ export const sceneLogic = kea<sceneLogicType>([
 
         return mapping
     }),
+
+    subscriptions(({ actions, values, cache }) => ({
+        title: (title) => {
+            const activeIndex = values.tabs.findIndex((t) => t.active)
+            if (activeIndex === -1) {
+                const { currentLocation } = router.values
+                actions.setTabs([
+                    {
+                        id: generateTabId(),
+                        active: true,
+                        pathname: currentLocation.pathname,
+                        search: currentLocation.search,
+                        hash: currentLocation.hash,
+                        title: title || 'Loading...',
+                    },
+                ])
+            } else {
+                if (!title || title === '...' || title === 'Loading...') {
+                    // When the tab is loading, don't flicker between the loaded title and the new one
+                    return
+                }
+                const newTabs = values.tabs.map((tab, i) => (i === activeIndex ? { ...tab, title } : tab))
+                actions.setTabs(newTabs)
+            }
+            // When the title changes, trigger a history REPLACE event to persist the new title in the browser
+            const { currentLocation } = router.values
+            router.actions.replace(currentLocation.pathname, currentLocation.search, currentLocation.hash)
+        },
+        tabs: () => {
+            const { tabIds } = values
+            for (const id of Object.keys(cache.mountedTabLogic)) {
+                if (!tabIds[id]) {
+                    const mountedTabLogic = cache.mountedTabLogic[id]
+                    if (mountedTabLogic) {
+                        try {
+                            mountedTabLogic.unmount()
+                        } catch (error) {
+                            console.error('Error unmounting tab logic:', error)
+                        }
+                    }
+                    delete cache.mountedTabLogic[id]
+                }
+            }
+        },
+    })),
 ])
